Allow `got` to select the correct `Agent` for the given request
When a cross-protocol redirect is encountered the agent needs to be
changed. There is no way to handle this at this level because the
redirect is handled internally in `got`. In
[`got@8.0.0`](https://github.com/sindresorhus/got/releases/tag/v8.0.0)
the ability to pass multiple agents (one for each protocol) was added to
solve this problem.

This also makes `http-https-agent` redundant as it is handled by `got`.